refactor(modals): merge BuyModal and SellModal into TradeModal
Closes #215. Replaces ~860 lines of duplicated JSX with a single
TradeModal component branching on a `mode` prop. Inline styles are
extracted to src/styles/trade-modal.css. Sell mode now seeds totalAmount
on mount so the price/total toggle works without first editing.
